// src/types/enum.d.ts
export type ButtonVariants =
  | "default"
  | "destructive"
  | "outline"
  | "secondary"
  | "ghost"
  | "link"

export type ButtonSizes = "default" | "icon" | "sm" | "lg"

export type Importance =
  | "Very Important"
  | "Important"
  | "Neutral"
  | "Unimportant"
  | "Very Unimportant"

export type FactorType = "Positive" | "Negative"